+++ tests/tracing/instrumentation/conftest.py
@@ -5,10 +5,8 @@ from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import SimpleSpanProcessor
 from opentelemetry.sdk.trace.export.in_memory_span_exporter import InMemorySpanExporter
 
-from rasa_sdk.executor import ActionExecutor, CollectingDispatcher
-from rasa_sdk.forms import ValidationAction, FormValidationAction
-from rasa_sdk.types import ActionCall, DomainDict
-from rasa_sdk import Tracker
+from rasa_sdk.executor import ActionExecutor
+from rasa_sdk.types import ActionCall
 
 
 @pytest.fixture(scope="session")
@@ -23,7 +21,7 @@ def span_exporter(tracer_provider: TracerProvider) -> InMemorySpanExporter:
     return exporter
 
 
-@pytest.fixture(scope="function")
+@pytest.fixture(scope="session")
 def previous_num_captured_spans(span_exporter: InMemorySpanExporter) -> int:
     captured_spans = span_exporter.get_finished_spans()  # type: ignore
     return len(captured_spans)
@@ -52,65 +50,3 @@ class MockActionExecutor(ActionExecutor):
         events: List[Dict[Text, Any]], messages: List[Dict[Text, Any]]
     ) -> None:
         pass
-
-
-class MockValidationAction(ValidationAction):
-    def __init__(self) -> None:
-        self.fail_if_undefined("run")
-
-    def fail_if_undefined(self, method_name: Text) -> None:
-        if not (
-            hasattr(self.__class__.__base__, method_name)
-            and callable(getattr(self.__class__.__base__, method_name))
-        ):
-            pytest.fail(
-                f"method '{method_name}' not found in {self.__class__.__base__}. "
-                f"This likely means the method was renamed, which means the "
-                f"instrumentation needs to be adapted!"
-            )
-
-    async def run(
-        self,
-        dispatcher: "CollectingDispatcher",
-        tracker: "Tracker",
-        domain: "DomainDict",
-    ) -> None:
-        pass
-
-    def name(self) -> Text:
-        return "mock_validation_action"
-
-    async def _extract_validation_events(
-        self,
-        dispatcher: "CollectingDispatcher",
-        tracker: "Tracker",
-        domain: "DomainDict",
-    ) -> None:
-        return tracker.events
-
-
-class MockFormValidationAction(FormValidationAction):
-    def __init__(self) -> None:
-        self.fail_if_undefined("run")
-
-    def fail_if_undefined(self, method_name: Text) -> None:
-        if not (
-            hasattr(self.__class__.__base__, method_name)
-            and callable(getattr(self.__class__.__base__, method_name))
-        ):
-            pytest.fail(
-                f"method '{method_name}' not found in {self.__class__.__base__}. "
-                f"This likely means the method was renamed, which means the "
-                f"instrumentation needs to be adapted!"
-            )
-
-    async def _extract_validation_events(
-        self,
-        dispatcher: "CollectingDispatcher",
-        tracker: "Tracker",
-        domain: "DomainDict",
-    ) -> None:
-        return tracker.events
-
-    def name(self) -> Text:
-        return "mock_form_validation_action"
